Time: 0 ms (100%), Space: 8.6 MB (69.81%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double myPow(double x, int n) {
-        long long binform = n;
+        long binform = n;
         if (n < 0){
             x = 1/x;
             binform = - binform;
